fix: usedStrorageUnits is float, cast it to int

+++ strato/api/organization.py
@@ -44,7 +44,7 @@ class OrganizationUsage:
     vectors: int
     styledMaps: int
     organizationMembers: int
-    usedStorageUnits: int
+    usedStorageUnits: float
 
 
 @dataclass
